requirements.txt
langchain==0.0.350
openai==1.3.0
pinecone-client==2.2.4
streamlit==1.28.0
python-dotenv==1.0.0
